requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
numpy
scikit-learn
opencv-python
matplotlib
google-generativeai
pandas
seaborn
Pillow